fix(viewer): resolve WASM build compilation errors
- Replace deprecated room_id_from_url() with current_room_id() in mode.rs
  (current_room_id already returns String, removed unwrap_or_else call)
- Cast TRPG_POLL_INTERVAL_MS (u64) to i32 for sleep_ms parameter in sse/client.rs

WASM build now completes successfully without errors.

+++ viewer/src/sse/client.rs
@@ -489,7 +489,7 @@ fn start_polling_loop(messages: Arc<Mutex<Vec<(String, String)>>>, active: Arc<A
                 }
             }
 
-            if sleep_ms(config::TRPG_POLL_INTERVAL_MS).await.is_err() {
+            if sleep_ms(config::TRPG_POLL_INTERVAL_MS as i32).await.is_err() {
                 break;
             }
         }
